Add a simple sinatra application negotiating GitHub OAuth in/out of Jenkins
This little app, as it stands right now, will share the same client_id and
client_secret as the configured Jenkins instance. That will allow it to log the
user in and authorize the OAuth application, before redirecting them "in" to
Jenkins to let the GitHub OAuth plugin set its own cookies.

This is wacky but it's the only way I can think to get us from
codevalet.io/github/authentication into rtyler.codevalet.io