Add some nginx/proxy configuration to allow docker-compose and kubernetes-based proxying
Getting environment variables into nginx is kind of a pain, and unfortunately on
recent versions of docker-compose it is impossible to get an IP address of a
linked container. Rather we need to jump through these god-awful hoops to run a
local resolver in the container in order to convert the string 'webapp' into an
IP address via dnsmasq and /etc/hosts.

BLECH